G1106_S06

+++ G1106_S05/app/src/main/java/ro/ase/ie/g1106_s05/activities/MovieActivity.java
@@ -125,7 +125,7 @@ public class MovieActivity extends AppCompatActivity {
                 Date release = null;
                 String poster = "";
                 Double budget = 0.0;
-                Double duration = 0.0;
+                Integer duration = 0;
                 Boolean recommended = null;
                 GenreEnum genre = null;
                 ParentalApprovalEnum status = null;
